Refactoring and test fixes
- Refactor updateResearchStudyWithClinicalStudy into its own module
- Fix tests
- Re-implement disabled test
