+++ spec/clinicaltrialsgov.spec.ts
@@ -1,7 +1,5 @@
-import { getContainedResource, ResearchStudy as ResearchStudyObj } from '../src/research-study';
-import { Address, FhirResource, Location, ResearchStudy, PlanDefinition } from 'fhir/r4';
+import { ResearchStudy } from 'fhir/r4';
 import * as ctg from '../src/clinicaltrialsgov';
-import { Study } from '../src/clinical-trials-gov';
 import fs from 'fs';
 import path from 'path';
 import nock from 'nock';
@@ -9,20 +7,11 @@ import { Volume } from 'memfs';
 
 // Trial missing summary, inclusion/exclusion criteria, phase and study type
 import trialMissing from './data/resource.json';
-import trialFilled from './data/complete_study.json';
 import { createClinicalStudy } from './support/clinicalstudy-factory';
 import { createResearchStudy } from './support/researchstudy-factory';
 import {
-  DateType,
-  DesignAllocation,
-  DesignMasking,
-  DesignTimePerspective,
-  InterventionalAssignment,
-  InterventionType,
-  ObservationalModel,
-  PrimaryPurpose,
-  Status,
-  StudyType
+  PagedStudies,
+  Study,
 } from '../src/ctg-api';
 
 function specFilePath(specFilePath: string): string {
@@ -127,11 +116,11 @@ describe('findNCTNumbers', () => {
 
 describe('parseStudyJson', () => {
   it('rejects invalid JSON', () => {
-    expect(ctg.parseStudyJson('true')).toThrowError('Invalid JSON object for Study: true');
+    expect(ctg.parseStudyJson('true')).toBeNull();
   });
   it('logs failures', () => {
     const log = jasmine.createSpy('log');
-    expect(ctg.parseStudyJson('not valid JSON', log)).toThrowError(/^Unable to parse JSON object/);
+    expect(ctg.parseStudyJson('not valid JSON', log)).toBeNull();
     expect(log).toHaveBeenCalled();
   });
 });
@@ -297,7 +286,7 @@ describe('ClinicalTrialsGovService', () => {
   const multipleEntriesDataDir = '/ctg-cache-multi';
   // The date used for when the cache was first created, 2021-02-01T12:00:00.000 (picked arbitrarily)
   const cacheStartTime = new Date(2021, 1, 1, 12, 0, 0, 0);
-  let study: ResearchStudy, nctID: ctg.NCTNumber;
+  let study: ResearchStudy;
   // Create our mock FS
   const cacheVol = Volume.fromNestedJSON({
     '/ctg-cache/data/NCT02513394.json': fs.readFileSync(specFilePath('NCT02513394.json'), { encoding: 'utf8' }),
@@ -320,8 +309,6 @@ describe('ClinicalTrialsGovService', () => {
     if (maybeNctID === null) {
       // This indicates a failure in test cases
       throw new Error('ResearchStudy has no NCT number');
-    } else {
-      nctID = maybeNctID;
     }
     // Create an empty directory
     cacheVol.mkdirSync('/ctg-cache-empty');
@@ -734,126 +721,85 @@ describe('ClinicalTrialsGovService', () => {
     let interceptor: nock.Interceptor;
     let downloader: ctg.ClinicalTrialsGovService;
     const nctIDs = ['NCT00000001', 'NCT00000002', 'NCT00000003'];
-    beforeEach(() => {
-      scope = nock('https://classic.clinicaltrials.gov');
-      interceptor = scope.get('/ct2/download_studies?term=' + nctIDs.join('+OR+'));
-      return ctg.createClinicalTrialsGovService(dataDirPath, { cleanInterval: 0, fs: cacheFS }).then((service) => {
-        downloader = service;
+    beforeEach(async () => {
+      scope = nock('https://clinicaltrials.gov');
+      interceptor = scope.get('/api/v2/studies?filter.ids=' + nctIDs.join(','));
+      // Need to intercept the writeFile method
+      spyOn(cacheFS, 'writeFile').and.callFake((_file, _data, _options, callback) => {
+        // For these, always pretend it succeeded
+        callback(null);
       });
+      downloader = await ctg.createClinicalTrialsGovService(dataDirPath, { cleanInterval: 0, fs: cacheFS });
     });
 
-    it('handles failures from https.get', () => {
+    it('handles failures from https.get', async () => {
       interceptor.replyWithError('Test error');
-      return expectAsync(downloader['downloadTrials'](nctIDs)).toBeRejectedWithError('Test error');
+      expect(await downloader['downloadTrials'](nctIDs)).toBeFalse();
     });
 
-    it('handles failure responses from the server', () => {
+    it('handles failure responses from the server', async () => {
       interceptor.reply(404, 'Unknown');
       // Pretend the middle entry exists
-      downloader['cache'].set(nctIDs[1], new ctg.CacheEntry(downloader, nctIDs[1] + '.xml', {}));
-      return expectAsync(
-        downloader['downloadTrials'](nctIDs).finally(() => {
-          expect(scope.isDone()).toBeTrue();
-        })
-      )
-        .toBeRejected()
-        .then(() => {
-          // Check to make sure the new cache entries do not still exist - the failure should remove them, but not the
-          // non-pending one
-          expect(downloader['cache'].has(nctIDs[0])).toBeFalse();
-          expect(downloader['cache'].has(nctIDs[2])).toBeFalse();
-        });
+      downloader['cache'].set(nctIDs[1], new ctg.CacheEntry(downloader, nctIDs[1] + '.json', {}));
+      expect(await downloader['downloadTrials'](nctIDs)).toBeFalse();
+      expect(scope.isDone()).toBeTrue();
+      // Check to make sure the new cache entries do not still exist - the failure should remove them, but not the
+      // non-pending one
+      expect(downloader['cache'].has(nctIDs[0])).toBeFalse();
+      expect(downloader['cache'].has(nctIDs[1])).toBeTrue();
+      expect(downloader['cache'].has(nctIDs[2])).toBeFalse();
     });
 
-    // it('creates cache entries', () => {
-    //   interceptor.reply(200, 'Unimportant', { 'Content-type': 'application/zip' });
-    //   // For this test, create an existing cache entry for one of the IDs
-    //   downloader['cache'].set(nctIDs[1], new ctg.CacheEntry(downloader, nctIDs[1] + '.xml', {}));
-    //   // Also mock the extraction process so it thinks everything is fine
-    //   downloader['extractResults'] = () => {
-    //     // Grab cache entries for our NCTs and say they've been resolved
-    //     for (const id of nctIDs) {
-    //       const entry = downloader['cache'].get(id);
-    //       expect(entry).toBeDefined();
-    //       if (entry) {
-    //         // Indicate that the entry is found
-    //         entry.found();
-    //       }
-    //     }
-    //     // Only mark entry 1 ready
-    //     const entry = downloader['cache'].get(nctIDs[1]);
-    //     if (entry) {
-    //       entry.ready();
-    //     }
-    //     return Promise.resolve();
-    //   };
-    //   return expectAsync(downloader['downloadTrials'](nctIDs))
-    //     .toBeResolved()
-    //     .then(() => {
-    //       // Should have created the two missing items which should still be pending as we mocked the extract process
-    //       let entry = downloader['cache'].get(nctIDs[0]);
-    //       expect(entry && entry.pending).toBeTrue();
-    //       entry = downloader['cache'].get(nctIDs[1]);
-    //       expect(entry && !entry.pending).toBeTrue();
-    //       entry = downloader['cache'].get(nctIDs[2]);
-    //       expect(entry && entry.pending).toBeTrue();
-    //     });
-    // });
+    it('creates cache entries', async () => {
+      interceptor.reply(
+        200,
+        JSON.stringify({
+          studies: nctIDs.map<Study>((id) => ({
+            protocolSection: {
+              identificationModule: {
+                nctId: id
+              }
+            }
+          }))
+        } as PagedStudies),
+        { 'Content-type': 'application/json' }
+      );
+      // For this test, create an existing cache entry for one of the IDs
+      downloader['cache'].set(nctIDs[1], new ctg.CacheEntry(downloader, nctIDs[1] + '.xml', {}));
+      expect(await downloader['downloadTrials'](nctIDs)).toBeTrue();
 
-    // it('extracts a ZIP', () => {
-    //   interceptor.reply(200, 'Unimportant', {
-    //     'Content-type': 'application/zip'
-    //   });
-    //   const spy = jasmine.createSpy('extractResults').and.callFake((): Promise<void> => {
-    //     // Grab cache entries for our NCTs and say they've been resolved
-    //     for (const id of nctIDs) {
-    //       const entry = downloader['cache'].get(id);
-    //       expect(entry).toBeDefined();
-    //       if (entry) {
-    //         entry.found();
-    //         entry.ready();
-    //       }
-    //     }
-    //     return Promise.resolve();
-    //   });
-    //   // Jam the spy in (method is protected, that's why it can't be created directly)
-    //   downloader['extractResults'] = spy;
-    //   return expectAsync(
-    //     downloader['downloadTrials'](nctIDs).finally(() => {
-    //       // Just check that it was called
-    //       expect(spy).toHaveBeenCalledTimes(1);
-    //       expect(scope.isDone()).toBeTrue();
-    //     })
-    //   ).toBeResolved();
-    // });
+      // Should have created the two missing items which should be resolved
+      let entry = downloader['cache'].get(nctIDs[0]);
+      expect(entry?.pending).toBeFalse();
+      entry = downloader['cache'].get(nctIDs[1]);
+      expect(entry?.pending).toBeFalse();
+      entry = downloader['cache'].get(nctIDs[2]);
+      expect(entry?.pending).toBeFalse();
+    });
 
-    // it('invalidates cache entries that were not found in the downloaded ZIP', () => {
-    //   // This is kind of complicated, but basically, we need to have it "create" the entries for the NCT IDs, but then
-    //   // have extractResults "not create" some of the entries.
-    //   interceptor.reply(200, 'Unimportant', {
-    //     'Content-type': 'application/zip'
-    //   });
-    //   const spy = jasmine.createSpy('extractResults').and.callFake((): Promise<void> => {
-    //     // For this test, we only mark one OK
-    //     const entry = downloader['cache'].get(nctIDs[1]);
-    //     expect(entry).toBeDefined();
-    //     if (entry) {
-    //       entry.found();
-    //       entry.ready();
-    //     }
-    //     return Promise.resolve();
-    //   });
-    //   // Jam the spy in (method is protected, that's why it can't be created directly)
-    //   downloader['extractResults'] = spy;
-    //   return expectAsync(downloader['downloadTrials'](nctIDs))
-    //     .toBeResolved()
-    //     .then(() => {
-    //       // The failed NCT IDs should be removed at this point
-    //       expect(downloader['cache'].has(nctIDs[0])).toBeFalse();
-    //       expect(downloader['cache'].has(nctIDs[1])).toBeTrue();
-    //       expect(downloader['cache'].has(nctIDs[2])).toBeFalse();
-    //     });
-    // });
+    it('invalidates cache entries that were not found in the results', async () => {
+      interceptor.reply(
+        200,
+        JSON.stringify({
+          // Only include NCT ID 1
+          studies: [
+            {
+              protocolSection: {
+                identificationModule: {
+                  nctId: nctIDs[1]
+                }
+              }
+            }
+          ]
+        } as PagedStudies),
+        { 'Content-type': 'application/json' }
+      );
+      expect(await downloader['downloadTrials'](nctIDs)).toBeTrue();
+      // The failed NCT IDs should be removed at this point
+      expect(downloader['cache'].has(nctIDs[0])).toBeFalse();
+      expect(downloader['cache'].has(nctIDs[1])).toBeTrue();
+      expect(downloader['cache'].has(nctIDs[2])).toBeFalse();
+    });
   });
 
   describe('#getCachedClinicalStudy', () => {
@@ -888,15 +834,15 @@ describe('ClinicalTrialsGovService', () => {
   describe('#ensureTrialsAvailable()', () => {
     // Most of these tests just pass through to downloadTrials
     let service: ctg.ClinicalTrialsGovService;
-    let downloadTrials: jasmine.Spy<(ids: string[]) => Promise<void>>;
+    let downloadTrials: jasmine.Spy<(ids: string[]) => Promise<boolean>>;
     beforeEach(() => {
       service = new ctg.ClinicalTrialsGovService(dataDirPath, { fs: cacheFS });
       // Can't directly spy on within TypeScript because downloadTrials is protected
-      const spy = jasmine.createSpy<(ids: string[]) => Promise<void>>('downloadTrials');
+      const spy = jasmine.createSpy<(ids: string[]) => Promise<boolean>>('downloadTrials');
       // Jam it in
       service['downloadTrials'] = spy;
       downloadTrials = spy;
-      downloadTrials.and.callFake(() => Promise.resolve());
+      downloadTrials.and.callFake(() => Promise.resolve(true));
     });
 
     it('excludes invalid NCT numbers in an array of strings', () => {
@@ -1041,704 +987,6 @@ describe('ClinicalTrialsGovService', () => {
       const testClinicalStudy = createClinicalStudy();
       service.updateResearchStudy(testResearchStudy, testClinicalStudy);
       // There's no really good way to verify this worked. For now, it not blowing up is good enough.
-    });
-  });
-
-  describe('filling out a partial trial', () => {
-    // Use the downloader to load the fixture data
-    let downloader: ctg.ClinicalTrialsGovService;
-    let updatedTrial: ResearchStudy;
-    let clinicalStudy: Study;
-    beforeAll(async function () {
-      downloader = new ctg.ClinicalTrialsGovService(dataDirPath, { cleanInterval: 0, fs: cacheFS });
-      await downloader.init();
-      // Cache should have been restored on init
-      const maybeStudy = await downloader.getCachedClinicalStudy(nctID);
-      if (maybeStudy === null) {
-        throw new Error('Unable to open study');
-      } else {
-        clinicalStudy = maybeStudy;
-      }
-      // Note this mutates study, which doesn't actually matter at present.
-      updatedTrial = ctg.updateResearchStudyWithClinicalStudy(study, clinicalStudy);
-    });
-
-    it('fills in inclusion criteria', () => {
-      expect(updatedTrial.enrollment).toBeDefined();
-      if (updatedTrial.enrollment) {
-        // Prove enrollment exists to TypeScript
-        expect(updatedTrial.enrollment.length).toBeGreaterThan(0);
-        expect(updatedTrial.enrollment[0].display).toBeDefined();
-      }
-    });
-
-    it('fills in phase', () => {
-      expect(updatedTrial.phase).toBeDefined();
-      if (updatedTrial.phase) expect(updatedTrial.phase.text).toBe('Phase 3');
-    });
-
-    it('fills in categories', () => {
-      expect(updatedTrial.category).toBeDefined();
-      if (updatedTrial.category) {
-        expect(updatedTrial.category.length).toEqual(5);
-        const categories = updatedTrial.category.map((item) => item.text);
-        expect(categories).toHaveSize(5);
-        expect(categories).toEqual(
-          jasmine.arrayContaining([
-            'Study Type: Interventional',
-            'Intervention Model: Parallel Assignment',
-            'Primary Purpose: Treatment',
-            'Masking: None (Open Label)',
-            'Allocation: Randomized'
-          ])
-        );
-      }
-    });
-
-    it('does not overwrite existing categories', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      researchStudy.category = [{ text: 'Study Type: Example' }];
-
-      ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          designModule: {
-            studyType: StudyType.INTERVENTIONAL,
-            designInfo: {
-              interventionModel: InterventionalAssignment.PARALLEL,
-              primaryPurpose: PrimaryPurpose.TREATMENT,
-              maskingInfo: {
-                masking: DesignMasking.NONE
-              },
-              allocation: DesignAllocation.RANDOMIZED,
-              timePerspective: DesignTimePerspective.OTHER,
-              observationalModel: ObservationalModel.CASE_CONTROL
-            }
-          }
-        }
-      });
-
-      expect(researchStudy.category).toBeDefined();
-      if (researchStudy.category) {
-        expect(researchStudy.category).toHaveSize(7);
-        const categories = researchStudy.category.map((item) => item.text);
-        expect(categories).toHaveSize(7);
-        expect(categories).toEqual(
-          jasmine.arrayContaining([
-            'Study Type: Example',
-            'Intervention Model: Parallel Assignment',
-            'Primary Purpose: Treatment',
-            'Masking: None (Open Label)',
-            'Allocation: Randomized',
-            'Time Perspective: Example',
-            'Observation Model: Something'
-          ])
-        );
-      }
-    });
-
-    it('will retain old categories if not part of standard study design', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      // Empty category but there is an object there for the sake of this test.
-      researchStudy.category = [{}];
-
-      ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          designModule: {
-            studyType: StudyType.INTERVENTIONAL
-          }
-        }
-      });
-
-      expect(researchStudy.category).toBeDefined();
-      if (researchStudy.category) {
-        expect(researchStudy.category).toHaveSize(2);
-      }
-    });
-
-    it('fills in arms', () => {
-      expect(updatedTrial.arm).toBeDefined();
-      if (updatedTrial.arm) {
-        expect(updatedTrial.arm).toHaveSize(2);
-        expect(updatedTrial.arm).toEqual(
-          jasmine.arrayContaining([
-            jasmine.objectContaining({
-              name: 'Arm A',
-              type: {
-                coding: jasmine.arrayContaining([{ code: 'Experimental', display: 'Experimental' }]),
-                text: 'Experimental'
-              },
-              description:
-                'Palbociclib at a dose of 125 mg orally once daily, Day 1 to Day 21 followed by 7 days off treatment in a 28-day cycle for a total duration of 2 years, in addition to standard adjuvant endocrine therapy for a duration of at least 5 years.'
-            }),
-            jasmine.objectContaining({
-              name: 'Arm B',
-              type: { coding: jasmine.arrayContaining([{ code: 'Other', display: 'Other' }]), text: 'Other' },
-              description: 'Standard adjuvant endocrine therapy for a duration of at least 5 years.'
-            })
-          ])
-        );
-      }
-    });
-
-    it('fills in protocol with interventions and arm references', () => {
-      expect(updatedTrial.protocol).toBeDefined();
-      if (updatedTrial.protocol) {
-        expect(updatedTrial.protocol).toHaveSize(3);
-        const references: PlanDefinition[] = [];
-        for (const plan of updatedTrial.protocol) {
-          if (plan.reference && plan.reference.length > 1) {
-            const intervention: PlanDefinition = getContainedResource(
-              updatedTrial,
-              plan.reference.substring(1)
-            ) as PlanDefinition;
-            if (intervention) references.push(intervention);
-          } else {
-            fail('PlanDefinition not defined for intervention');
-          }
-        }
-
-        try {
-          const titles = references.map((item) => item.title);
-          const types = references.map((item) => (item.type ? item.type.text : null));
-          const subjects = references.map((item) =>
-            item.subjectCodeableConcept ? item.subjectCodeableConcept.text : null
-          );
-
-          expect(titles).toEqual(
-            jasmine.arrayContaining([
-              'Palbociclib',
-              'Standard Adjuvant Endocrine Therapy',
-              'Standard Adjuvant Endocrine Therapy'
-            ])
-          );
-          expect(types).toEqual(jasmine.arrayContaining(['Drug', 'Drug', 'Drug']));
-          expect(subjects).toEqual(jasmine.arrayContaining(['Arm A', 'Arm A', 'Arm B']));
-        } catch (err) {
-          fail(err);
-        }
-      }
-    });
-
-    it('fills in interventions even without arms', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          armsInterventionsModule: {
-            interventions: [
-              {
-                type: InterventionType.BEHAVIORAL,
-                name: 'Name',
-                description: 'Description',
-                otherNames: ['Other name']
-              }
-            ]
-          }
-        }
-      });
-
-      expect(result.protocol).toBeDefined();
-      expect(result.protocol).toHaveSize(1);
-
-      if (result.protocol && result.protocol.length > 0) {
-        if (result.protocol[0].reference && result.protocol[0].reference.length > 1) {
-          const intervention: PlanDefinition = getContainedResource(
-            result,
-            result.protocol[0].reference.substring(1)
-          ) as PlanDefinition;
-          expect(intervention).toEqual(
-            jasmine.objectContaining({
-              resourceType: 'PlanDefinition',
-              status: 'unknown',
-              description: 'Description',
-              title: 'Name',
-              subtitle: 'Other name',
-              type: { text: 'Behavioral' }
-            })
-          );
-        }
-      }
-    });
-
-    it('fills in interventions with description and subtitle', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          armsInterventionsModule: {
-            interventions: [
-              {
-                type: InterventionType.BEHAVIORAL,
-                name: 'Name',
-                description: 'Description',
-                otherNames: ['Other name'],
-                armGroupLabels: ['Arm']
-              }
-            ]
-          }
-        }
-      });
-
-      expect(result.protocol).toBeDefined();
-      expect(result.protocol).toHaveSize(1);
-
-      if (result.protocol && result.protocol.length > 0) {
-        if (result.protocol[0].reference && result.protocol[0].reference.length > 1) {
-          const intervention: PlanDefinition = getContainedResource(
-            result,
-            result.protocol[0].reference.substring(1)
-          ) as PlanDefinition;
-          expect(intervention).toEqual(
-            jasmine.objectContaining({
-              resourceType: 'PlanDefinition',
-              status: 'unknown',
-              description: 'Description',
-              title: 'Name',
-              subtitle: 'Other name',
-              type: { text: 'Behavioral' },
-              subjectCodeableConcept: { text: 'Arm' }
-            })
-          );
-        }
-      }
-    });
-
-    it('fills in period', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          statusModule: {
-            startDateStruct: {
-              date: '2023-01',
-              type: DateType.ACTUAL
-            },
-            completionDateStruct: {
-              date: '2023-02',
-              type: DateType.ACTUAL
-            }
-          }
-        }
-      });
-
-      expect(result.period).toBeDefined();
-      if (result.period) {
-        expect(result.period.start).toBeDefined();
-        expect(result.period.end).toBeDefined();
-
-        expect(result.period.start).toEqual(new Date('January 2022').toISOString());
-        expect(result.period.end).toEqual(new Date('February 2023').toISOString());
-      }
-    });
-
-    it('fills in start of period even without end', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          statusModule: {
-            startDateStruct: {
-              date: '2023-01',
-              type: DateType.ACTUAL
-            }
-          }
-        }
-      });
-
-      expect(result.period).toBeDefined();
-      if (result.period) {
-        expect(result.period.start).toBeDefined();
-        expect(result.period.end).not.toBeDefined();
-
-        expect(result.period.start).toEqual(new Date('January 2022').toISOString());
-      }
-    });
-
-    it('fills in end of period even without start', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          statusModule: {
-            completionDateStruct: {
-              date: '2023-02',
-              type: DateType.ACTUAL
-            }
-          }
-        }
-      });
-
-      expect(result.period).toBeDefined();
-      if (result.period) {
-        expect(result.period.start).not.toBeDefined();
-        expect(result.period.end).toBeDefined();
-
-        expect(result.period.end).toEqual(new Date('February 2023').toISOString());
-      }
-    });
-
-    it('does not fill in period if not a real date', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          statusModule: {
-            startDateStruct: {
-              date: 'Not real',
-              type: DateType.ACTUAL
-            },
-            completionDateStruct: {
-              date: 'Not real',
-              type: DateType.ACTUAL
-            }
-          }
-        }
-      });
-
-      expect(result.period).not.toBeDefined();
-    });
-
-    it('fills in description', () => {
-      expect(updatedTrial.description).toBeDefined();
-    });
-
-    it('fills out the status', () => {
-      const actual = ctg.updateResearchStudyWithClinicalStudy(
-        { resourceType: 'ResearchStudy', status: 'active' },
-        {
-          protocolSection: {
-            statusModule: {
-              lastKnownStatus: Status.AVAILABLE
-            }
-          }
-        }
-      );
-      expect(actual.status).toEqual('completed');
-    });
-
-    it('leaves status alone if unavailable', () => {
-      const actual = ctg.updateResearchStudyWithClinicalStudy(
-        { resourceType: 'ResearchStudy', status: 'active' },
-        {
-          // Lie about types
-          protocolSection: {
-            statusModule: {
-              lastKnownStatus: 'something invalid' as unknown as Status
-            }
-          }
-        }
-      );
-      // It shouldn't have changed it, because it can't
-      expect(actual.status).toEqual('active');
-    });
-
-    it('fills out conditions', () => {
-      const actual = ctg.updateResearchStudyWithClinicalStudy(
-        { resourceType: 'ResearchStudy', status: 'active' },
-        {
-          protocolSection: {
-            conditionsModule: {
-              conditions: ['Condition 1', 'Condition 2']
-            }
-          }
-        }
-      );
-      expect(actual.condition).toBeDefined();
-      if (actual.condition) {
-        expect(actual.condition.length).toEqual(2);
-        expect(actual.condition[0].text).toEqual('Condition 1');
-        expect(actual.condition[1].text).toEqual('Condition 2');
-      }
-    });
-
-    it('fills in contact', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          contactsLocationsModule: {
-            centralContacts: [
-              {
-                name: 'First Middle Last, MD',
-                phone: '1112223333',
-                email: 'email@example.com'
-              },
-              {
-                name: 'First2 Middle2 Last2, DO',
-                phone: '1234567890',
-                email: 'email2@example.com'
-              }
-            ]
-          }
-        }
-      });
-
-      expect(result.contact).toBeDefined();
-      if (result.contact) {
-        expect(result.contact).toHaveSize(2);
-        expect(result.contact).toEqual(
-          jasmine.arrayContaining([
-            jasmine.objectContaining({
-              name: 'First Middle Last, MD',
-              telecom: [
-                { system: 'email', value: 'email@example.com', use: 'work' },
-                { system: 'phone', value: '1112223333', use: 'work' }
-              ]
-            }),
-            jasmine.objectContaining({
-              name: 'First2 Middle2 Last2, DO',
-              telecom: [
-                { system: 'email', value: 'email2@example.com', use: 'work' },
-                { system: 'phone', value: '1234567890', use: 'work' }
-              ]
-            })
-          ])
-        );
-      }
-    });
-
-    it('fills in contacts even with missing information', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          contactsLocationsModule: {
-            centralContacts: [
-              {
-                name: 'First Last',
-                email: 'email@example.com'
-              },
-              {
-                name: 'Middle2',
-                phone: '1234567890'
-              }
-            ]
-          }
-        }
-      });
-
-      expect(result.contact).toBeDefined();
-      if (result.contact) {
-        expect(result.contact).toHaveSize(2);
-        expect(result.contact).toEqual(
-          jasmine.arrayContaining([
-            jasmine.objectContaining({
-              name: 'First Last',
-              telecom: [{ system: 'email', value: 'email@example.com', use: 'work' }]
-            }),
-            jasmine.objectContaining({
-              name: 'Middle2',
-              telecom: [{ system: 'phone', value: '1234567890', use: 'work' }]
-            })
-          ])
-        );
-      }
-    });
-
-    it('does not overwrite site data', () => {
-      const researchStudy = new ResearchStudyObj('id');
-      const location = researchStudy.addSite('Example');
-      const result = ctg.updateResearchStudyWithClinicalStudy(researchStudy, {
-        protocolSection: {
-          contactsLocationsModule: {
-            locations: [
-              {
-                facility: 'Facility'
-              }
-            ]
-          }
-        }
-      });
-      expect(result.site).toBeDefined();
-      const sites = result.site;
-      if (sites) {
-        expect(sites.length).toEqual(1);
-        if (sites[0]) {
-          expect(sites[0].reference).toEqual('#' + location.id);
-          if (location.id) {
-            const actualLocation = getContainedResource(result, location.id);
-            expect(actualLocation).not.toBeNull();
-            if (actualLocation) {
-              expect(actualLocation.resourceType).toEqual('Location');
-              expect((actualLocation as Location).name).toEqual('Example');
-            }
-          } else {
-            fail('location.id not defined');
-          }
-        } else {
-          fail('sites[0] undefined');
-        }
-      }
-    });
-
-    it('does not alter a filled out trial', () => {
-      // Clone the trial in the dumbest but also most sensible way
-      const exampleStudy: ResearchStudy = JSON.parse(JSON.stringify(trialFilled));
-      ctg.updateResearchStudyWithClinicalStudy(exampleStudy, clinicalStudy);
-      // Currently active gets overwritten intentioanlly, so set the example
-      // back to its original value even if it changed
-      // (Note that the "as" *should* verify that the underlying JSON value is
-      // in fact valid at compile time. I think.)
-      exampleStudy.status = trialFilled.status as ResearchStudy['status'];
-      // Nothing should have changed
-      expect(exampleStudy).toEqual(trialFilled as ResearchStudy);
-    });
-
-    function expectTelecom(location: Location, type: 'phone' | 'email', expectedValue: string | null) {
-      // Look through the telecoms
-      // If the expected value is null, telecom must be defined, otherwise it
-      // may be empty
-      if (expectedValue !== null) expect(location.telecom).toBeDefined();
-      if (location.telecom) {
-        // If we're expecting a telecom we're expecting it to appear exactly once
-        let found = 0;
-        for (const telecom of location.telecom) {
-          if (telecom.system === type) {
-            found++;
-            if (found > 1) {
-              fail(`Found an extra ${type}`);
-            }
-            if (expectedValue === null) {
-              // If null, it wasn't expected at all
-              fail(`Expected no ${type}, but one was found`);
-            } else {
-              expect(telecom.use).toEqual('work');
-              expect(telecom.value).toEqual(expectedValue);
-            }
-          }
-        }
-        if (expectedValue !== null && found === 0) {
-          fail(`Expected one ${type}, not found`);
-        }
-      }
-    }
-
-    function expectLocation(
-      resource: FhirResource,
-      expectedName?: string,
-      expectedPhone?: string,
-      expectedEmail?: string,
-      expectedAddress?: Address
-    ) {
-      if (resource.resourceType === 'Location') {
-        const location = resource as Location;
-        if (expectedName) {
-          expect(location.name).toEqual(expectedName);
-        } else {
-          expect(location.name).not.toBeDefined();
-        }
-        expectTelecom(location, 'phone', expectedPhone || null);
-        expectTelecom(location, 'email', expectedEmail || null);
-        if (expectedAddress) {
-          expect(location.address).toBeDefined();
-          expect(location.address).toEqual(expectedAddress);
-        } else {
-          expect(location.address).not.toBeDefined();
-        }
-      } else {
-        fail(`Expected Location, got ${resource.resourceType}`);
-      }
-    }
-
-    it('fills out sites as expected', () => {
-      const result = ctg.updateResearchStudyWithClinicalStudy(
-        { resourceType: 'ResearchStudy', status: 'active' },
-        {
-          protocolSection: {
-            contactsLocationsModule: {
-              locations: [
-                // Everything in location is optional, so this is valid:
-                {},
-                {
-                  facility: 'No Details'
-                },
-                {
-                  facility: 'Only Email',
-                  contacts: [
-                    {
-                      email: 'email@example.com'
-                    }
-                  ]
-                },
-                {
-                  facility: 'Only Phone',
-                  contacts: [
-                    {
-                      phone: '781-555-0100'
-                    }
-                  ]
-                },
-                {
-                  facility: 'Phone and Email',
-                  contacts: [
-                    {
-                      email: 'hasemail@example.com',
-                      phone: '781-555-0101'
-                    }
-                  ]
-                },
-                {
-                  facility: 'Only Address',
-                  city: 'Bedford',
-                  state: 'MA',
-                  country: 'US',
-                  zip: '01730',
-                  contacts: [
-                    {
-                      email: 'email@example.com'
-                    }
-                  ]
-                }
-              ]
-            }
-          }
-        }
-      );
-      // Sites should be filled out
-      expect(result.site).toBeDefined();
-      if (result.site) {
-        expect(result.site.length).toEqual(6);
-      }
-      // Make sure each individual site was created properly - they will be contained resources and should be in order
-      expect(result.contained).toBeDefined();
-      if (result.contained) {
-        // Both 0 and 1 should be empty
-        expectLocation(result.contained[0]);
-        expectLocation(result.contained[1]);
-        expectLocation(result.contained[2], 'Only Email', undefined, 'email@example.com');
-        expectLocation(result.contained[3], 'Only Phone', '781-555-0100');
-        expectLocation(result.contained[4], 'Phone and Email', '781-555-0101', 'hasemail@example.com');
-        expectLocation(result.contained[5], 'Only Address', undefined, undefined, {
-          use: 'work',
-          city: 'Bedford',
-          state: 'MA',
-          postalCode: '01730',
-          country: 'US'
-        });
-      }
-    });
-
-    function expectEmptyResearchStudy(researchStudy: ResearchStudy): void {
-      // Technically this is just checking fields updateResearchStudyWithClinicalStudy may change
-      expect(researchStudy.contained).withContext('contained').not.toBeDefined();
-      expect(researchStudy.enrollment).withContext('enrollment').not.toBeDefined();
-      expect(researchStudy.description).withContext('description').not.toBeDefined();
-      expect(researchStudy.phase).withContext('phase').not.toBeDefined();
-      expect(researchStudy.category).withContext('category').not.toBeDefined();
-      expect(researchStudy.status).toEqual('active');
-      expect(researchStudy.condition).withContext('condition').not.toBeDefined();
-      expect(researchStudy.site).withContext('site').not.toBeDefined();
-    }
-
-    it("handles JSON with missing data (doesn't crash)", () => {
-      let researchStudy: ResearchStudy;
-      // According to the schema, literally everything is optional, so an empty object should "work"
-      researchStudy = ctg.updateResearchStudyWithClinicalStudy(new ResearchStudyObj('id'), {});
-      // Expect nothing to have changed
-      expectEmptyResearchStudy(researchStudy);
-      // Some partial JSON
-      researchStudy = ctg.updateResearchStudyWithClinicalStudy(new ResearchStudyObj('id'), {
-        protocolSection: {
-          eligibilityModule: {
-            genderBased: false,
-            minimumAge: '18 years',
-            maximumAge: 'N/A'
-          }
-        }
-      });
-      expectEmptyResearchStudy(researchStudy);
     });
   });
 });
